services/workload_inference/pyproject.toml
[project]
name = "workload-inference"
version="0.1.0"
requires-python = ">=3.12"

dependencies = [
    "pyzmq",
    "scikit-learn",
    "pandas",
    "matplotlib",
    "numpy",
    "debugpy",
    "PyYaml",
    "PyQt6",
    "requests"
]

[dependency-groups]
dev = [
    "pytest",
    "black",
    "mypy",
    "ruff"
]

[project.scripts]
workload_inference = "workload_inference.main:main"
visualize_task = "workload_inference.visualizer.replay:main"
generate_data = "workload_inference.test_data:main"

[tool.black]
line_length = 88

[tool.mypy]
python_version = "3.14"
disable_error_code = ["import-untyped"]
check_untyped_defs = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py314"
[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # Logging format
    "G",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"